【实战系列 12】登录进Game逻辑

+++ Unity/Assets/Scripts/Model/Generate/Client/Message/OuterMessage_C_10001.cs
@@ -1608,7 +1608,7 @@ namespace ET
 		public string Message { get; set; }
 
 		[MemoryPackOrder(3)]
-		public int MyUnitId { get; set; }
+		public long MyUnitId { get; set; }
 
 		public override void Dispose() 
 		{
